requirements.txt
beautifulsoup4==4.12.3
python-fasthtml==0.4.2
requests==2.32.3

